* Use fixed_size_simd<int, Size> as IndexType for all ABIs * Define integral AVX compares only when AVX2 is available * Refactor binary operators to handle (mixed) fixed_size_simd(_mask) and   Simd(Mask)Array better (or at all) * Modify Simd(Mask)Array operations to return fixed_size_simd(_mask) * Refactor some enable_ifs to move from default arg to template   parameter list. This is especially important to work around   constructor inheritance bugs in GCC. * Make sure fixed_size unconditionally passes via the stack (to solve   potential ABI breakage on linking TUs compiled with different -march   flags). * Disallow implicit narrowing conversions on binary operators (e.g.   `simd + long`)

+++ tests/mask.cpp
@@ -240,7 +240,7 @@ TEST_TYPES(Vec, testZero, AllVectors) /*{{{*/
     typedef typename Vec::Mask Mask;
     typedef typename Vec::IndexType I;
 
-    for (size_t cut = 0; cut < Vec::Size; ++cut) {
+    for (int cut = 0; cut < int(Vec::Size); ++cut) {
         const Mask mask = Vc::simd_cast<Mask>(I([](int n) { return n; }) < cut);
         //std::cout << mask << std::endl;
